requirements.txt
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
gunicorn
bleach
pillow
Flask-SSLify
flask-sqlalchemy
psycopg2-binary